pyproject.toml
[project]
name = "tabq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.6.0",
    "transformers==4.49.0",
    "opencv-python==4.7.0.72",
    "accelerate>=1.4.0",
    "pillow>=11.1.0",
    "hf-transfer>=0.1.9",
    "tensorboardx>=2.6.2.2",
    "python-levenshtein>=0.27.1",
    "qwen-vl-utils>=0.0.10",
    "opencv-python-headless>=4.11.0.86",
    "torchvision>=0.21.0",
    "apted>=1.0.3",
    "distance>=0.1.3",
    "lxml>=5.3.2",
    "scikit-learn>=1.6.1",
    "gradio>=5.47.0",
    "nltk>=3.9.1",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[dependency-groups]
dev = [
    "jsonlines>=4.0.0",
    "pymupdf>=1.25.3",
]
